refactor: moved product list types to the domain folder
