+++ src/actions/list/listServices.ts
@@ -2,8 +2,8 @@
 import { z } from "zod";
 
 import { TFilters, TListItem } from "@/domains/store/productList/types";
+import { TListSort } from "@/domains/store/productList/types/";
 import { db } from "@/shared/lib/db";
-import { TListSort } from "@/types/list";
 import { TProductPath } from "@/types/product";
 
 const ValidateSort = z.object({
