+++ src/domains/store/productList/constants/index.ts
@@ -1,4 +1,4 @@
-import { TListSort } from "@/types/list";
+import { TListSort } from "@/domains/store/productList/types/";
 import { TDropDown } from "@/types/uiElements";
 
 import { TFilters } from "../types";
