--- src/types/list.ts
@@ -1,11 +1,0 @@
-export type TListSort = {
-  sortName: "id" | "price" | "name";
-  sortType: "asc" | "desc";
-};
-
-export type TPageStatus =
-  | "pageLoading"
-  | "filterLoading"
-  | "filledProductList"
-  | "filterHasNoProduct"
-  | "categoryHasNoProduct";
